Upgraded chain length to a 64 bit integer so that chain lengths > 2.1 billion can be done. This addresses issue 540. Runs OK, accepts long chain lengths but not actually tested a run of > 2.1 billion.

+++ trunk/src/dr/evomodel/tree/Convergence.java
@@ -117,7 +117,7 @@ public class Convergence {
         return taxa;
     }
 
-    public double log(int state) {
+    public double log(long state) {
 
         if (logEvery <= 0 || ((state % logEvery) == 0)) {
